requirements.txt
requests
pyaudio
pyttsx3
python-dotenv
customtkinter
Pillow
pygame
openai